Updates score to be unoptimizable
Villager: 1 for team win, 2 for guessing each mafia
Mafia: 3 for losing, 2 points for not being guessed as mafia (one point deduction per guess on you)

Previously, it was more advantageous for the mafia to play well and get 3 points for not being guessed than it was to lose.  I think that the primary goal of the mafia should be to lose.  This variation also only works with 1 mafia on each time. We'll probably have to find a way to fit the 1 mafia game mode as well.